Rename the project and add an authentification
WIP since all the security check are not done in the EntryController & all the tests are not done yet

+++ WebApp/Controllers/EntriesController.cs
@@ -2,26 +2,41 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using WebApp.Data;
 using WebApp.Models;
+using Microsoft.AspNetCore.Authorization;
 
 namespace WebApp.Controllers
 {
     public class EntriesController : Controller
     {
-        private readonly WebAppContext _context;
+        private readonly ApplicationDbContext _context;
+        private readonly SignInManager<IdentityUser> _signInManager;
+        private readonly UserManager<IdentityUser> _userManager;
 
-        public EntriesController(WebAppContext context)
+        public EntriesController(ApplicationDbContext context, SignInManager<IdentityUser> signInManager, UserManager<IdentityUser> userManager)
         {
             _context = context;
+            _signInManager = signInManager;
+            _userManager = userManager;
         }
 
         // GET: Entries
         public async Task<IActionResult> Index(string title, string journal)
         {
+
+            if (!_signInManager.IsSignedIn(User))
+            {
+                return Redirect("/Identity/Account/Login");
+            }
+
+            var userId = _userManager.GetUserId(User);
             var entries = from e in _context.Entry
+                          where e.UserId == userId
                           select e;
 
             if (!string.IsNullOrEmpty(title))
@@ -34,6 +49,7 @@ namespace WebApp.Controllers
             }
 
             var journals = from e in _context.Entry
+                           where e.UserId == userId
                            select e.Journal;
 
             var model = new EntriesIndexViewModel
@@ -41,7 +57,7 @@ namespace WebApp.Controllers
                 TitleFilter = title,
                 JournalFilter = journal,
                 Journals = await journals.Distinct().ToListAsync(),
-                Entries = await entries.OrderByDescending(e => e.Journal).ThenBy(e=> e.Id).ToListAsync()
+                Entries = await entries.OrderByDescending(e => e.Journal).ThenBy(e => e.Id).ToListAsync()
             };
             return View(model);
         }
@@ -49,13 +65,17 @@ namespace WebApp.Controllers
         // GET: Entries/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            if (!_signInManager.IsSignedIn(User))
+            {
+                return Redirect("/Identity/Account/Login");
+            }
             if (id == null)
             {
                 return NotFound();
             }
 
             var entry = await _context.Entry
-                .FirstOrDefaultAsync(m => m.Id == id);
+                .FirstOrDefaultAsync(m => m.Id == id && m.UserId == _userManager.GetUserId(User));
             if (entry == null)
             {
                 return NotFound();
@@ -67,6 +87,10 @@ namespace WebApp.Controllers
         // GET: Entries/Create
         public IActionResult Create()
         {
+            if (!_signInManager.IsSignedIn(User))
+            {
+                return Redirect("/Identity/Account/Login");
+            }
             return View();
         }
 
@@ -77,6 +101,11 @@ namespace WebApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Journal,Title,Pages")] Entry entry)
         {
+            if (!_signInManager.IsSignedIn(User))
+            {
+                return Redirect("/Identity/Account/Login");
+            }
+            entry.UserId = _userManager.GetUserId(User);
             if (ModelState.IsValid)
             {
                 _context.Add(entry);
@@ -89,13 +118,17 @@ namespace WebApp.Controllers
         // GET: Entries/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            if (!_signInManager.IsSignedIn(User))
+            {
+                return Redirect("/Identity/Account/Login");
+            }
             if (id == null)
             {
                 return NotFound();
             }
 
             var entry = await _context.Entry.FindAsync(id);
-            if (entry == null)
+            if (entry == null || entry.UserId != _userManager.GetUserId(User))
             {
                 return NotFound();
             }
@@ -109,6 +142,11 @@ namespace WebApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Journal,Title,Pages")] Entry entry)
         {
+            // TODO Check if the ID related to the item to edit is owned by the currently connected user.
+            if (!_signInManager.IsSignedIn(User))
+            {
+                return Redirect("/Identity/Account/Login");
+            }
             if (id != entry.Id)
             {
                 return NotFound();
@@ -140,13 +178,17 @@ namespace WebApp.Controllers
         // GET: Entries/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (!_signInManager.IsSignedIn(User))
+            {
+                return Redirect("/Identity/Account/Login");
+            }
             if (id == null)
             {
                 return NotFound();
             }
 
             var entry = await _context.Entry
-                .FirstOrDefaultAsync(m => m.Id == id);
+                .FirstOrDefaultAsync(m => m.Id == id && m.UserId == _userManager.GetUserId(User));
             if (entry == null)
             {
                 return NotFound();
@@ -160,7 +202,11 @@ namespace WebApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var entry = await _context.Entry.FindAsync(id);
+            if (!_signInManager.IsSignedIn(User))
+            {
+                return Redirect("/Identity/Account/Login");
+            }
+            var entry = await _context.Entry.FirstOrDefaultAsync(m => m.Id == id && m.UserId == _userManager.GetUserId(User));
             _context.Entry.Remove(entry);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
